requirements.txt
coreapi==2.3.3
coreschema==0.0.4
django-filter==22.1
djangorestframework==3.14.0
drf-yasg==1.21.4
Pillow==9.3.0

